build: Use exported targets for libjpeg-turbo and bump min to 2.1
libjpeg-turbo has exported cmake configs since its version 2.1.
Relying on this simplifies the build and allows us to completely
remove the troublesome Findlibjpeg-turbo.cmake. This implies raising
the floor for libjpeg-turbo support to 2.1 (the current release is
3.0) which was released in April 2021 (beta in Nov 2020).

That's cutting it a little close (we usually support dependencies
going back at least 3 years at a minimum, and often significantly
longer).  But since this is an optional dependency anyway and
simplifies the build, I'm tempted to do so.

Signed-off-by: Larry Gritz <lg@larrygritz.com>