fix: to_OpenCV fails for ImageCache-backed images
We were using parallel_convert_image but not realizing that asking for
the IB::localpixels() would give us nullptr for an IC-backed image.

Instead, "wrap" the cv::Mat with an IB and then use IBA::copy().  This
handles both in-memory and IC-backed ImageBuf's, and also has many
other advantages (such as format conversion) we may wish to take
advantage of in the future.

Signed-off-by: Larry Gritz <lg@larrygritz.com>