Deprecate Sysutil::physical_concurrency() (#4034)
It is a implemented by simply wrapping boost::physical_concurrency(). So
this is one more step in eliminating boost from our requirements.

Until it's fully removed, just make it be a synonym for
hardware_concurrency().

We don't seem to use physical_concurrency anywhere, and I'm not sure
what it's good for. The underlying boost impementation just falls back
to hardware_concurrency on many platforms or circumstances.

Note for the record that Sysutil::hardware_concurrency() is also
redundant, and merely wraps the std::thread function of the same name.

Signed-off-by: Larry Gritz <lg@larrygritz.com>