fix: oiiotool --autocc and color config inventory cleanup
The important bug fix is in oiiotool image input when autocc is used,
where we reversed the sense of a test and did the autocc-upon-input if
the file's color space was equivalent to the scene_linear space
(pointlessly), and skipped the conversion if they were different (oh
no, big bug!).

Along the way:

* Add missing try/catch around OCIO call that should have had it.

* Some very ugly SPI-specific recognize-by-name color space heuristics.
  I hate this, sorry, but it solves a really important problem for me.

* A bunch of additional debugging messages during color space inventory,
  enabled only when debugging, so nobody should see these but me.

* A couple places where we canonicalize the spelling of "oiio:ColorSpace".

Note that there is still an issue with the color space classification
using OCIO 2.3+'s identification of built-in equivalents by
name. These are shockingly expensive. I will return to this later.

Signed-off-by: Larry Gritz <lg@larrygritz.com>