test: improve color management test in imagebufalgo_test (#4063)
Add a few more assertion checks and error messages in imagebufalgo_test.

I observe that if somebody runs the testsuite with `$OCIO` set to a
config that doesn't contain the required color space names for this
test, it will fail. As a backstop, use the default built-in config if
the initial request fails. That will always give a working processor
between these spaces if OCIO >= 2.2 is being used.

Also one more tiny SPI-specific enhancement to hacky identification of
nonconforming color space names in our old configs.

Signed-off-by: Larry Gritz <lg@larrygritz.com>