Fix link to openexr test images (#4080)
## Description

Minimal change to update link to OpenEXR's new Test Images page. I poked
around a bit to find a "friendlier" variant of the link but could find
none.

Note that the section containing this link implies that folks need to
download these files themselves and should be "placed in a sibling
directory to oiio-testimages". The problem is that not even OpenEXR
lists the repository for the test images nor does the page linked above
offer a download.

Should we list out the openexr-images repo address too and say to just
clone that?

Fixes #4042

---------

Signed-off-by: Jesse Yurkovich <jesse.y@gmail.com>