feat(oiiotool): oiiotool --cryptomatte-colors
This patch adds a new oiiotool command: `--cryptomatte-colors`, which
takes the name of a cryptomatte set of channels, and produces a
color-coded matte in which each ID gets a distinct color in the image.
This can be useful for visualizing the matte, among other things.

Maybe this is the first cryptomatte related functionality of many?

Signed-off-by: Larry Gritz <lg@larrygritz.com>