ci: Repair Sonar scanner analysis (#4097)
This broke some time back, probably partly related to our change in
repository home. This seems to get it mostly working? The reports are
still not showing up, but I think the solution might be to make sure
this is checked into master, so we need to merge this before assessing
whether there is more work to be done in this area.

Signed-off-by: Larry Gritz <lg@larrygritz.com>