fix(int): Simplify test and make static analysis happy
The prior test was redundant. Since a few lines above, if B was an
image but A was not, we swapped to ensure that A was always an
image. So here, we can just test A to know for sure that at least one
of them is an image. That makes the test here simpler, but it also
makes it more clear to static analysis that from this point forward, A
can't be nullptr.

Signed-off-by: Larry Gritz <lg@larrygritz.com>