WIP api(IBA): replace filter name or raw Filter2D* with FilterOpt
Let's consider this a "design review", or a prototype, as I'm really
not sure if this is the right design.

Inspired by Issue 4102.

Here's the gist:

There are a number of ImageBufAlgo functions that need to do
filtering, and so need a variety of additional parameters like
filtername and width, but sometimes other things, and they don't all
keep up wth each other in terms of their parameters. Also, in addition
to name/width, they also have additional varieties that take a
`Filter2D*`, so that doubles the number of entry points. It's a little
confusing. So this PR proposes the following:

* A new helper class FilterOpt that collects the different filtering
  parameters -- name and width, or Filter2D*, wrap mode, and other
  params TBD.
* The IBA functions that do filtering get cleaned up and take a
  FilterOpt& parameter. They have fewer varieties now. The old versions
  are scheduled for deprecation and moved to the unobtrusive end of
  imagebufalgo.h. So this doesn't break source compatibility for now,
  we'll more formally deprecate with warnings later, then for 3.0 remove
  the old call signatures entirely.
* filter.h cleaned up a bit to give Filter1D and Filter2D a new
  create_shared static method that returns a shared_ptr.

I think this is a step to making a more uniform and simple API for the
IBA functions that take filtering controls. Here are a few examples
of how the calls would look.

Old calls:

    // specify filter by name and width
    r = IBA::resize (src, "lanczos3", 6.0f, roi);

    // if we have a Filter2D* already:
    r = IBA::resize (src, filtptr, 6.0f, roi);

New calls, various varieties (all one API signature, I'm essentially
showing off the various FilterOpt constructors):

    r = IBA::resize (src, {}, roi); // default filter choice
    r = IBA::resize (src, FilterOpt("lanczos3", 6.0f), roi);  // explicit
    r = IBA::resize (src, "lanczos3", roi); // use name and its default width
    r = IBA::resize (src, filtptr, roi); // constructs from Filter2D*

    r = IBA::resize (src, { .name = "lanczos3", .width = 6.0f }, roi);
    // ^^ This fancy syntax needs C++20 officially, but apparently works
    // in gcc 8+ clang 10+, MSVS 19.21+, Apple clang 12+, and icx 2021.1,
    // all versions we'll have as our minimums relatively shortly.
    // Maybe this is the future, as the closest approximation to named
    // parameter passing we can get in C++?

Does this seem appealing as an API? Maybe some people will love
passing a block of related params as a struct, others will hate it?

What happens when we want to add additional options? Can we only
do that when it's ok to break ABI and can never backport new fields?
Should we use a PIMPL setup to hide the internals of FilerOpt so it's
not part of the ABI (at the expense of pointer indirections to access
the fields)? Embed a "version" number in the struct and on the library
side use that to cast to the right one for the version it represents?

I expect we'll probably iterate on this design quite a bit before we
get consensus. There is no rush on this, it's for master / 2.6.

Signed-off-by: Larry Gritz <lg@larrygritz.com>