feat: query/print build information
New OIIO global attributes (read only):

  - "build:platform" retrieves the OS and CPU type that this build of
    OIIO is fof (e.g., "Linux/x86_64").
  - "build:compiler" retrieves the compiler name and version used to
    build OIIO itself (e.g. "gcc 9.3").
  - "build:dependencies" retrieves a semicolon-separated list of
    dependency libraries -- both for image format support (like
    libtiff) but also general (boost, TBB, Python, fmt, OCIO).
  - "build:simd" is a new (preferred) synonym for the SIMD and other
    hardware capabilities selected at build time. The oldname,
    "oiio:simd" is hereby softly deprecated as confusing.

`oiiotool --buildinfo` is a new command that prints this information.

Example output:

```
$ oiiotool --buildinfo

OIIO 2.6.0.2spi | MacOS/x86_64
    Build compiler: Apple clang 15.0 | C++17/201703
    HW features enabled at build: sse2,sse3,ssse3,sse41,sse42
Dependencies: OpenEXR 3.2.1, LIBTIFF Version 4.6.0, jpeg-turbo
    3.0.1/jp80, dcmtk 3.6.8, FFMpeg 6.0 (Lavf60.16.100), gif_lib 5.2.1,
    libheif 1.17.6, OpenJpeg 2.5.0, null 1.0, OpenVDB 11.0.0abi11, libpng
    1.6.40, Ptex 2.4, libraw 0.21.2-Release, Webp 1.3.2, Boost 1.83,
    OpenColorIO 2.3.1, NO Python!, TBB 2021.11.0, fmt 10.2.1
```

This is all a convenience utility to make it easy to ask users to run
a simple command that will tell us a lot of information relevant to
resolving their issues in cases where it's related to some of their
build-time compiler or dependencies.

Also cleaned up the "bug report issue template" to clarify the
instructions and specifically ask people to run this command and
include the information with any bug report.

Signed-off-by: Larry Gritz <lg@larrygritz.com>