fix: certain int->float type conversions in TypeDesc/ParamValueList (#4132)
We had all along omitted an important 'int' conversion case for an
internal to_floats() helper function that's used by the
TypeDesc::convert_type utility. Which is in turn used by
ParamValue::get_float() and ParamValueList::get_float().

All of these would fail to return the proper float if the underlying
data was an int, despite these functions saying that they would have no
trouble returning the float equivalent of int data when asked.

I added a test that verifies that it's wrking.

Signed-off-by: Larry Gritz <lg@larrygritz.com>
Signed-off-by: Peter Kovář <peter.kovar@reflexion.tv>