int: span and range checking enhancements (#4125)
* New header span_util.h with span-related helpers:
  - make_span/make_cspan helps for converting vectors & arrays to spans
when passing to templated functions.
  - spancpy, spanset, spanzero are memory-safe, range-checking
substitutions for memcpy, memset when working with spans.
* Add range check DASSERT to span access in debug mode. There were a
couple spots where we were taking the address of an out-of-range value
that needed to be adjusted to use a better idiom.
* ParamValue: extra OIIO_DASSERT range check for get() method.
* Simplify constructors by replacing `&x[0]` with using `data()`, which
is always well-defined.
* Fix typos in span.h comments

The range checking imposes no extra cost for release/optimized builds.
But it might help us for CI analysis and any time we're debugging.

Signed-off-by: Larry Gritz <lg@larrygritz.com>
Signed-off-by: Peter Kovář <peter.kovar@reflexion.tv>